feat: 전력 그리드 시스템이 서버에게 전송한 온도 정보 처리 [SOCKET_SERVER33]

+++ smart_city_socket_server/protocol/protocol_request_packt.h
@@ -8,7 +8,7 @@ struct _protocol_request_packt
     int total_length;
     int target_command;
     int sub_command;
-    int data[0];
+    float data[0];
     //char ip_addr[16];
     //int session_id;
 };
